Refactor entity sync: type-specific packets & uint32_t IDs
Refactored world entity networking to use explicit, type-specific packets for players, enemies, and objects, replacing generic "Entity" packets. All entity IDs (players, objects, enemies) are now uint32_t throughout the codebase. Updated event system, packet registry, and world managers to handle new packet types and containers. PhysicsManager and entity classes updated for new types and ID sizes. Peer management now uses uint32_t IDs. Added new packet headers for Player/Enemy/Object spawn, update, and destroy. Improves type safety, extensibility, and scalability of world state sync.

+++ EchoDungeon/Game/World/Entities/Object.cpp
@@ -4,7 +4,7 @@ Object::Object()
     : asset_id(""), type(ObjectType::MODEL), id(65534) {
 }
 
-Object::Object(uint16_t id, const std::string& asset_id, ObjectType type)
+Object::Object(uint32_t id, const std::string& asset_id, ObjectType type)
     : asset_id(asset_id), type(type), id(id) {
 }
